local make install working?
