+++ docs/building_optvl_locally.md
@@ -13,11 +13,11 @@ The `pyproject.toml` should hold all the information pip needs to use the meson 
 To modify aspects of the build, such as the compilation flags, see the `meson.build` file. 
 
 ## Using Make
-This method is more convenient for quick compilation during editing. The steps are
+This method is more convenient for quick compilation during editing, but only works on Mac and Linux. The steps are
 
 1. Copy a config file from `config/defaults` to the config directory, i.e. `cp  config/defaults/config.LINUX_GFORTRAN.mk config/config.mk`
 2. Create the shared library file with `make`
-3. Install the python wrapper with `python setup_deprecated.py develop`
+3. Install the python wrapper with `pip install -e ./dev_make_instal`
 
 After modifying the Fortran code only `make` needs to be run again. 
 If you only modify the python code then there is no need to rerun anything.
